Move RotationConfig Constraints into seperate file
